don't track Promise owners, simplify map_inbound
The kernel doesn't actually care which vat allocated a given Promise: it just
needs to know where to route messages aimed at it. So don't track the owner
in the PromiseTable.

And map_inbound() doesn't really need to know who owns the object/promise it
is mapping. If the target vat owns it, the index will already be in the
clist (with a positive vat-side ID). If not, it won't. So remove the
owner_of() checks and use map_inbound() all the time, instead of sometimes
doing get_inbound() and asserting that it was already present.